tests/requirements.txt
# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel test execution

# Kubernetes testing
kubernetes==28.1.0
pyyaml==6.0.1

# API testing
requests==2.31.0
httpx==0.25.1

# Database testing
psycopg2-binary==2.9.9
redis==5.0.1

# Kafka testing
confluent-kafka==2.3.0

# Data testing
pandas==2.1.3
pyarrow==14.0.1

# Performance testing
locust==2.18.3  # Load testing
pytest-benchmark==4.0.0

# Coverage and reporting
coverage[toml]==7.3.2
codecov==2.1.13

# Mocking
pytest-mock==3.12.0
responses==0.24.1

# Additional utilities
faker==20.1.0  # Test data generation
freezegun==1.4.0  # Time mocking


